doDeepSleep: changed the datatype of time_us to uint64_t, according to ESP8266 Arduino core 2.4.1
As a result it is possible to set durations longer than ~71min.

+++ src/Homie.hpp
@@ -68,7 +68,7 @@ class HomieClass {
   //FIXME implement on ESP32
   #elif defined(ESP8266)
   static void prepareToSleep();
-  static void doDeepSleep(uint32_t time_us = 0, RFMode mode = RF_DEFAULT);
+  static void doDeepSleep(uint64_t time_us = 0, RFMode mode = RF_DEFAULT);
   #endif // ESP32
 
  private:
